+++ R/plot_dvconc.R
@@ -23,7 +23,7 @@
 #'
 #' @examples
 #'data_sad_pd <- dplyr::filter(data_sad, CMT ==3)
-#'data <- df_addn(dplyr::mutate(data_sad_pd, Dose = DOSE), grp_var = Dose, sep = "mg")
+#'data <- dplyr::mutate(data_sad_pd, Dose = var_addn(DOSE, ID, sep = "mg"))
 #'plot_dvconc(data, dv_var = ODV, idv_var = CONC, col_var = Dose, col_trend = FALSE)
 #'