+++ R/plot_dvtime.R
@@ -59,7 +59,7 @@
 #'
 #' @examples
 #'data_sad_pk <- dplyr::filter(data_sad, CMT %in% c(1,2))
-#'data <- df_addn(dplyr::mutate(data_sad_pk, Dose = DOSE), grp_var = Dose, sep = "mg")
+#'data <- dplyr::mutate(data_sad_pk, Dose = var_addn(DOSE, ID, sep = "mg"))
 #'plot_dvtime(data, dv_var = ODV, cent = "median", col_var = Dose)
 #'