+++ tests/testthat/test-plot_dvtime_dual.R
@@ -2,19 +2,19 @@
 
 ##Test Output
 test_that("Output is a `ggplot` plot object", {
-  data <- df_addn(dplyr::mutate(data_sad, Dose = DOSE), grp_var = Dose, sep = "mg")
+  data <- dplyr::mutate(data_sad, Dose = var_addn(DOSE, ID, sep = "mg"))
   expect_s3_class(plot_dvtime_dual(data, dv_var1 = "ODV", dv_var2 = "ODV", col_var = "Dose"),
                class = "ggplot")
 })
 
 test_that("Output contains two patchwork panels", {
-  data <- df_addn(dplyr::mutate(data_sad, Dose = DOSE), grp_var = Dose, sep = "mg")
+  data <- dplyr::mutate(data_sad, Dose = var_addn(DOSE, ID, sep = "mg"))
   p <- plot_dvtime_dual(data, dv_var1 = "ODV", dv_var2 = "ODV", col_var = "Dose")
   expect_equal(length(p$patches$plots) + 1, 2)
 })
 
 test_that("Custom ylab1 is applied to the first panel", {
-  data <- df_addn(dplyr::mutate(data_sad, Dose = DOSE), grp_var = Dose, sep = "mg")
+  data <- dplyr::mutate(data_sad, Dose = var_addn(DOSE, ID, sep = "mg"))
   p <- plot_dvtime_dual(data, dv_var1 = "ODV", dv_var2 = "ODV", col_var = "Dose",
                         ylab1 = "PK", ylab2 = "PD")
   # patchwork: first plot goes to patches$plots, base plot is the second
@@ -22,7 +22,7 @@ test_that("Custom ylab1 is applied to the first panel", {
 })
 
 test_that("Custom ylab2 is applied to the second panel", {
-  data <- df_addn(dplyr::mutate(data_sad, Dose = DOSE), grp_var = Dose, sep = "mg")
+  data <- dplyr::mutate(data_sad, Dose = var_addn(DOSE, ID, sep = "mg"))
   p <- plot_dvtime_dual(data, dv_var1 = "ODV", dv_var2 = "ODV", col_var = "Dose",
                         ylab1 = "PK", ylab2 = "PD")
   # patchwork: base plot is the last added (dv2)
@@ -30,14 +30,14 @@ test_that("Custom ylab2 is applied to the second panel", {
 })
 
 test_that("onelegend = TRUE collects guides", {
-  data <- df_addn(dplyr::mutate(data_sad, Dose = DOSE), grp_var = Dose, sep = "mg")
+  data <- dplyr::mutate(data_sad, Dose = var_addn(DOSE, ID, sep = "mg"))
   p <- plot_dvtime_dual(data, dv_var1 = "ODV", dv_var2 = "ODV", col_var = "Dose",
                         onelegend = TRUE)
   expect_equal(p$patches$layout$guides, "collect")
 })
 
 test_that("onelegend = FALSE does not collect guides", {
-  data <- df_addn(dplyr::mutate(data_sad, Dose = DOSE), grp_var = Dose, sep = "mg")
+  data <- dplyr::mutate(data_sad, Dose = var_addn(DOSE, ID, sep = "mg"))
   p <- plot_dvtime_dual(data, dv_var1 = "ODV", dv_var2 = "ODV", col_var = "Dose",
                         onelegend = FALSE)
   expect_false(identical(p$patches$layout$guides, "collect"))
@@ -55,7 +55,7 @@ test_that("Error if dvid_var does not exist in data", {
 
 ##Test NSE Bare Names
 test_that("plot_dvtime_dual accepts bare names", {
-  data <- df_addn(dplyr::mutate(data_sad, Dose = DOSE), grp_var = Dose, sep = "mg")
+  data <- dplyr::mutate(data_sad, Dose = var_addn(DOSE, ID, sep = "mg"))
   expect_s3_class(plot_dvtime_dual(data, dv_var1 = ODV, dv_var2 = ODV, col_var = Dose),
                   "ggplot")
 })